0000050 Исправлено не корректное чтение tiledata.mul. Привел количество ячеек для мульти и статики в соответствие с оригиналом.

+++ OrionUO/OrionUO.h
@@ -171,10 +171,10 @@ public:
 
 
 	//Получить флаги ландшафта
-	uint GetLandFlags(const ushort &id);
+	__int64 GetLandFlags(const ushort &id);
 
 	//Получить флаги статики
-	uint GetStaticFlags(const ushort &id);
+	__int64 GetStaticFlags(const ushort &id);
 
 	//Получить цвет света
 	ushort GetLightColor(const ushort &id) { return m_StaticDataIndex[id].LightColor; }
